# docs/openapi.yaml
...
components:
  schemas:
...
    AdminSummary:
      description: One admin of the community.
      properties:
        designation:
          anyOf:
          - type: string
          - type: 'null'
          title: Designation
        displayRole:
          title: Displayrole
          type: string
        email:
          anyOf:
          - type: string
          - type: 'null'
          title: Email
        flat:
          anyOf:
          - type: string
          - type: 'null'
          title: Flat
        id:
          title: Id
          type: string
        joinedAt:
          format: date-time
          title: Joinedat
          type: string
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        phone:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone
        profileId:
          title: Profileid
          type: string
        role:
          title: Role
          type: string
        status:
          title: Status
          type: string
        unitId:
          anyOf:
          - type: string
          - type: 'null'
          title: Unitid
      required:
      - id
      - profileId
      - role
      - displayRole
      - status
      - joinedAt
      title: AdminSummary
      type: object
...
    ApprovedRegistration:
      description: 'Result of approving a registration request.


        Approval mints an **invitation**, it does not create an active account: the

        invite token is a mandatory second factor. ``link`` and ``code`` are shown to

        the admin exactly once and are never recoverable -- only their digests are

        stored.'
      properties:
        code:
          title: Code
          type: string
        expiresAt:
          format: date-time
          title: Expiresat
          type: string
        flat:
          title: Flat
          type: string
        invitationId:
          title: Invitationid
          type: string
        link:
          title: Link
          type: string
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        phone:
          title: Phone
          type: string
        requestId:
          title: Requestid
          type: string
      required:
      - requestId
      - invitationId
      - link
      - code
      - phone
      - flat
      - expiresAt
      title: ApprovedRegistration
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    MessageResult:
      description: Acknowledgement for an action with nothing meaningful to return.
      properties:
        message:
          title: Message
          type: string
      required:
      - message
      title: MessageResult
      type: object
    ModuleCollection:
      description: 'The ten modules and the two counts worth showing above them.


        ``enabledWithoutBackend`` is the honest number: modules switched on that

        nothing in the product implements.'
      properties:
        enabledCount:
          title: Enabledcount
          type: integer
        enabledWithoutBackend:
          title: Enabledwithoutbackend
          type: integer
        items:
          items:
            $ref: '#/components/schemas/ModuleSummary'
          title: Items
          type: array
        total:
          title: Total
          type: integer
      required:
      - items
      - total
      - enabledCount
      - enabledWithoutBackend
      title: ModuleCollection
      type: object
    ModuleSummary:
      description: 'One feature module, as the community has it set.


        ``backendStatus`` is the field that makes this list worth reading: six of the

        ten modules have no backend at all, so an admin can otherwise switch on

        Parking Management and be given no hint that nothing will happen.


        ``isDefault`` is true when nobody has ever set this key for this community,

        which is how an eleventh module added to the catalogue reads before anyone

        touches it.'
      properties:
        backendNote:
          title: Backendnote
          type: string
        backendStatus:
          title: Backendstatus
          type: string
        backendStatusLabel:
          title: Backendstatuslabel
          type: string
        defaultEnabled:
          title: Defaultenabled
          type: boolean
        description:
          title: Description
          type: string
        enabled:
          title: Enabled
          type: boolean
        isDefault:
          title: Isdefault
          type: boolean
        key:
          title: Key
          type: string
        name:
          title: Name
          type: string
        sortOrder:
          title: Sortorder
          type: integer
        updatedAt:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Updatedat
        updatedBy:
          anyOf:
          - type: string
          - type: 'null'
          title: Updatedby
      required:
      - key
      - name
      - description
      - enabled
      - isDefault
      - defaultEnabled
      - backendStatus
      - backendStatusLabel
      - backendNote
      - sortOrder
      title: ModuleSummary
      type: object
    ModuleToggleRequest:
      description: 'Turn one module on or off.


        One key at a time, because two admins toggling two different modules in the

        same minute must not undo each other -- which is exactly what a

        whole-set write does.'
      properties:
        enabled:
          title: Enabled
          type: boolean
      required:
      - enabled
      title: ModuleToggleRequest
      type: object
...
    Page_AdminSummary_:
      properties:
        hasMore:
          title: Hasmore
          type: boolean
        items:
          items:
            $ref: '#/components/schemas/AdminSummary'
          title: Items
          type: array
        page:
          title: Page
          type: integer
        pageSize:
          title: Pagesize
          type: integer
        total:
          title: Total
          type: integer
      required:
      - items
      - total
      - page
      - pageSize
      - hasMore
      title: Page[AdminSummary]
      type: object
...
    Page_RegistrationRequestSummary_:
      properties:
        hasMore:
          title: Hasmore
          type: boolean
        items:
          items:
            $ref: '#/components/schemas/RegistrationRequestSummary'
          title: Items
          type: array
        page:
          title: Page
          type: integer
        pageSize:
          title: Pagesize
          type: integer
        total:
          title: Total
          type: integer
      required:
      - items
      - total
      - page
      - pageSize
      - hasMore
      title: Page[RegistrationRequestSummary]
      type: object
...
    RegistrationRequestSummary:
      description: One pending self-signup awaiting admin review.
      properties:
        date:
          title: Date
          type: string
        email:
          anyOf:
          - type: string
          - type: 'null'
          title: Email
        flat:
          title: Flat
          type: string
        id:
          title: Id
          type: string
        name:
          title: Name
          type: string
        phone:
          title: Phone
          type: string
        status:
          title: Status
          type: string
        submittedAt:
          format: date-time
          title: Submittedat
          type: string
        tower:
          anyOf:
          - type: string
          - type: 'null'
          title: Tower
      required:
      - id
      - name
      - phone
      - flat
      - status
      - date
      - submittedAt
      title: RegistrationRequestSummary
      type: object
...
    RejectRegistrationRequest:
      description: Body for rejecting a registration request.
      properties:
        reason:
          anyOf:
          - maxLength: 500
            type: string
          - type: 'null'
          title: Reason
      title: RejectRegistrationRequest
      type: object
    ReplaceModulesRequest:
      description: 'Set the whole module set from the list of keys that should be on.


        The shape the onboarding wizard already produces: ``enabledModules`` is an

        array of the enabled keys and every other key is off by omission. An empty

        list is legitimate and means every module off; a missing field is a `422`,

        because a caller who forgot the field would otherwise disable everything.'
      properties:
        moduleKeys:
          items:
            type: string
          maxItems: 50
          title: Modulekeys
          type: array
      required:
      - moduleKeys
      title: ReplaceModulesRequest
      type: object
...
    SettingsSnapshot:
      description: 'Everything the settings screen needs, in one response.


        ``hasSavedSettings`` is false for a community that has never saved: the

        values are then defaults rather than choices, and a screen that cannot tell

        the two apart will show "Asia/Kolkata" as though somebody picked it.'
      properties:
        billing:
          $ref: '#/components/schemas/BillingToggles'
        community:
          $ref: '#/components/schemas/CommunityProfile'
        hasSavedSettings:
          title: Hassavedsettings
          type: boolean
        modules:
          $ref: '#/components/schemas/ModuleCollection'
        preferences:
          $ref: '#/components/schemas/PreferenceSettings'
        updatedAt:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Updatedat
        updatedBy:
          anyOf:
          - type: string
          - type: 'null'
          title: Updatedby
        version:
          title: Version
          type: integer
      required:
      - community
      - preferences
      - billing
      - modules
      - hasSavedSettings
      - version
      title: SettingsSnapshot
      type: object
...
    UpdateResidentRequest:
      description: 'Editable fields on a resident.


        Every field is optional: this is a PATCH, and omitting a field leaves it

        unchanged. Explicitly sending ``null`` clears it.'
      properties:
        designation:
          anyOf:
          - maxLength: 100
            type: string
          - type: 'null'
          title: Designation
        email:
          anyOf:
          - maxLength: 320
            type: string
          - type: 'null'
          title: Email
        name:
          anyOf:
          - maxLength: 200
            type: string
          - type: 'null'
          title: Name
        phone:
          anyOf:
          - maxLength: 32
            type: string
          - type: 'null'
          title: Phone
      title: UpdateResidentRequest
      type: object
    UpdateSettingsRequest:
      description: 'Patch the community preferences. Omitted fields are left unchanged.


        ``unitLabelSingular`` distinguishes omitted from null: sending null clears

        the override and goes back to deriving the label from the community type,

        which is not the same as leaving it alone. Key presence is what separates

        them, matching ``defaultMaintenanceAmount`` on the billing patch.'
      properties:
        inviteTtlHours:
          anyOf:
          - maximum: 720.0
            minimum: 1.0
            type: integer
          - type: 'null'
          title: Invitettlhours
        noticeSmsBroadcastEnabled:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Noticesmsbroadcastenabled
        requireVisitorPreapproval:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Requirevisitorpreapproval
        timezone:
          anyOf:
          - maxLength: 64
            minLength: 3
            type: string
          - type: 'null'
          title: Timezone
        unitLabelSingular:
          anyOf:
          - maxLength: 24
            type: string
          - type: 'null'
          title: Unitlabelsingular
        visitorCodeTtlMinutes:
          anyOf:
          - maximum: 1440.0
            minimum: 5.0
            type: integer
          - type: 'null'
          title: Visitorcodettlminutes
      title: UpdateSettingsRequest
      type: object
...
  securitySchemes:
    HTTPBearer:
      scheme: bearer
      type: http
info:
  summary: Backend for the HomeBandhu residential community platform.
  title: HomeBandhu API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /api/v1/admins:
    get:
      description: 'Every admin of the caller''s community.


        Returned in the standard page envelope for consistency, but not paginated --

        an association has a committee, not a directory.'
      operationId: list_admins_api_v1_admins_get
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_AdminSummary_'
          description: Successful Response
      security:
      - HTTPBearer: []
      summary: List admins
      tags:
      - people
...
  /api/v1/registrations:
    get:
      description: Self-signup requests awaiting review, newest first.
      operationId: list_registrations_api_v1_registrations_get
      parameters:
      - description: Which requests to return.
        in: query
        name: status
        required: false
        schema:
          default: pending
          description: Which requests to return.
          pattern: ^(pending|approved|rejected)$
          title: Status
          type: string
      - in: query
        name: page
        required: false
        schema:
          default: 1
          minimum: 1
          title: Page
          type: integer
      - in: query
        name: pageSize
        required: false
        schema:
          default: 20
          maximum: 100
          minimum: 1
          title: Pagesize
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_RegistrationRequestSummary_'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: List registration requests
      tags:
      - people
  /api/v1/registrations/{request_id}/approve:
    post:
      description: 'Approve a request and mint the applicant''s invitation, atomically.


        **Approval does not create an active account.** It issues an invite the

        applicant redeems, because the invite token is a mandatory second factor.

        The returned ``link`` and ``code`` are shown once and are not recoverable.


        201 rather than 200: this creates an invitation.'
      operationId: approve_registration_api_v1_registrations__request_id__approve_post
      parameters:
      - in: path
        name: request_id
        required: true
        schema:
          title: Request Id
          type: string
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovedRegistration'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Approve a registration request
      tags:
      - people
  /api/v1/registrations/{request_id}/reject:
    post:
      description: Reject a pending request. The applicant may apply again.
      operationId: reject_registration_api_v1_registrations__request_id__reject_post
      parameters:
      - in: path
        name: request_id
        required: true
        schema:
          title: Request Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/RejectRegistrationRequest'
              - type: 'null'
              title: Body
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResult'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Reject a registration request
      tags:
      - people
...
  /api/v1/residents/{membership_id}:
    delete:
      description: 'Deactivate a resident and end their residency.


        **This is not a delete.** Complaints, invoices and payments reference the

        membership; removing the row would either cascade them away or fail outright.

        The member stops appearing in active lists and the flat reads as vacant.


        Returns 409 if an admin tries to remove their own membership -- there is no

        recovery path from locking a community out of its own dashboard.'
      operationId: remove_resident_api_v1_residents__membership_id__delete
      parameters:
      - in: path
        name: membership_id
        required: true
        schema:
          title: Membership Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResult'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Remove a resident
      tags:
      - people
    patch:
      description: Patch a resident's editable fields. Omitted fields are left unchanged.
      operationId: update_resident_api_v1_residents__membership_id__patch
      parameters:
      - description: The membership id, not the profile id.
        in: path
        name: membership_id
        required: true
        schema:
          description: The membership id, not the profile id.
          title: Membership Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateResidentRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResult'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Update a resident
      tags:
      - people
  /api/v1/settings:
    get:
      description: 'Everything the settings screen needs, in one request.


        ``hasSavedSettings`` is ``false`` for a community that has never saved. The

        values are then defaults rather than choices, and rendering the two the same

        way tells an admin they picked a timezone they have never seen.'
      operationId: get_settings_api_v1_settings_get
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettingsSnapshot'
          description: Successful Response
      security:
      - HTTPBearer: []
      summary: The community settings snapshot
      tags:
      - settings
    put:
      description: 'Patch the community preferences. Omitted fields are left unchanged.


        ``unitLabelSingular: null`` clears the override and goes back to deriving the

        word from the community type, which is not what omitting the field does.'
      operationId: update_settings_api_v1_settings_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSettingsRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettingsSnapshot'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Patch the community preferences
      tags:
      - settings
  /api/v1/settings/modules:
    get:
      description: 'The ten feature modules and this community''s setting for each.


        **Any authenticated role**, unlike the rest of this file: if module state ever

        gates navigation then every shell needs it, and a resident learning the

        marketplace is off discloses nothing.


        ``backendStatus`` is the field worth rendering. Six of the ten modules have no

        backend at all, so without it an admin can switch on Parking Management and

        be given no hint that nothing will happen. ``enabledWithoutBackend`` is that

        count.'
      operationId: list_modules_api_v1_settings_modules_get
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModuleCollection'
          description: Successful Response
      security:
      - HTTPBearer: []
      summary: List the feature modules
      tags:
      - settings
    put:
      description: 'Set the whole module set from the list of keys that should be on.


        The shape the onboarding wizard already produces: ``enabledModules`` is the

        array of enabled keys, and every other key is off by omission. Every key is

        validated before anything is written, so one typo does not leave a community

        half-configured.


        An empty array is legitimate and turns everything off. A missing field is a

        `422` rather than the same thing, because a caller who forgot it would

        otherwise disable the whole product by accident.'
      operationId: replace_modules_api_v1_settings_modules_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReplaceModulesRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModuleCollection'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Replace the whole module set
      tags:
      - settings
  /api/v1/settings/modules/{moduleKey}:
    patch:
      description: 'Turn one module on or off.


        One key at a time so two admins toggling two different modules in the same

        minute do not undo each other -- which is what a whole-set write does.'
      operationId: set_module_api_v1_settings_modules__moduleKey__patch
      parameters:
      - description: A catalogue key such as `amenities-booking`. Unknown keys 404.
        in: path
        name: moduleKey
        required: true
        schema:
          description: A catalogue key such as `amenities-booking`. Unknown keys 404.
          maxLength: 64
          title: Modulekey
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModuleToggleRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModuleSummary'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Toggle one feature module
      tags:
      - settings
...
servers:
- description: Local development
  url: http://localhost:8000
- description: Deployed
  url: https://{host}
  variables:
    host:
      default: api.homebandhu.example
      description: API host
...